refactor: remove the need for `~` in the route files name
