+++ dashboard/src/routes/__root.tsx
