+++ dashboard/src/routes/index.tsx
