+++ dashboard/src/routes/tree/$treeId/index.tsx
