+++ dashboard/tsr.config.json
@@ -1,5 +1,4 @@
 {
-  "routeFilePrefix": "~",
   "routesDirectory": "./src/routes",
   "generatedRouteTree": "./src/routeTree.gen.ts"
 }
